Student led session proposal
I asked Rachel about doing something like this last module and she thought it was a good idea. The field day bumped around the student session schedules right after so I didn't have a chance to get it in. I'm a bit trepidatious about it but I'd like to try it and see how it goes.