Update module-10-12-15.markdown
This falls under the approved top-level categories of Comptuer Hardware and DevOps. Topics will include computer process architecture and benchmarking, networking and bottlenecks, cloud computing, and other portions of end-to-end web application delivery. It will use stories from my experience with IT providers, as both buyer and seller, as examples of how to ask smart questions and work with technical sales people.