Update module-11-30-15.markdown
Jason Pilz and myself would like to lead a session on how to get started using Sketch for design. We feel it will benefit all students interested in web design and branding. Our goal is to split the hour with the first half covering basics and the second using the tools to create a logo.